Create README.md
This case study aims to build a predictive analytics model for an education company named X Education, which sells online courses to industry professionals. Here we will try to optimize the sales operation of company by finding out the the most prominent leads out of large dataset of people who interacted with the company website.